Updated app.py; moved db_connection imports
