+++ eathelp/__init__.py
@@ -4,6 +4,8 @@ from flask_sqlalchemy import SQLAlchemy
 from credentials import get_db_credentials
 from eathelp.api import api
 
+db = SQLAlchemy()
+
 # Based on http://flask.pocoo.org/docs/1.0/tutorial/factory/#the-application-factory
 # Modified to use Flask SQLAlchemy
 def create_app(test_config=None):
@@ -31,5 +33,3 @@ def create_app(test_config=None):
     db.init_app(app)
     app.register_blueprint(api.blueprint)
     return app
-
-db = SQLAlchemy()
