+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import find_packages, setup
 
 setup(
     name="eathelp",
-    version="0.1.0",
+    version="0.0.1",
     packages=find_packages(),
     include_package_data=True,
     zip_safe=False,
